pyproject.toml
[tool.poetry]
name = "ai-api-lint"
version = "0.1.0"
description = "AI-friendly API linter for OpenAPI specifications"
authors = ["SonAIengine <sonsj97@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "ai_api_lint"}]

[tool.poetry.scripts]
ai-api-lint = "ai_api_lint.cli:main"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
pyyaml = {version = ">=6.0", optional = true}

[tool.poetry.extras]
yaml = ["pyyaml"]

[tool.poetry.group.dev.dependencies]
ruff = ">=0.8"
pytest = ">=7.0"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.per-file-ignores]
"ai_api_lint/rules/*.py" = ["N801"]  # AI001_RuleName convention

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
